move create_self_signed from CaSecret to Secret class
